requirements.txt
Django<=1.4.5
django-recaptcha
